refactor(app): move deck layer blocklist
