+++ app/src/local-resources/deck_configuration/getStandardDeckViewLayerBlockList.ts
