+++ app/src/local-resources/deck_configuration/index.ts
@@ -1,0 +1,1 @@
+export * from './getStandardDeckViewLayerBlockList'
